Refactoring and fix dep cycle
