+++ packages/grid/_modules_/grid/components/column-header-menu-icon.tsx
@@ -1,6 +1,5 @@
 import * as React from 'react';
 import IconButton from '@material-ui/core/IconButton';
-import { useGridState } from '../hooks/features/core/useGridState';
 import { useIcons } from '../hooks/utils/useIcons';
 import { ApiContext } from './api-context';
 import { ColDef } from '../models/colDef/colDef';
@@ -14,16 +13,10 @@ export const ColumnHeaderMenuIcon: React.FC<ColumnHeaderFilterIconProps> = React
     const icons = useIcons();
     const icon = React.createElement(icons.ColumnMenu!, {});
     const apiRef = React.useContext(ApiContext);
-    const [, setGridState, forceUpdate] = useGridState(apiRef!);
 
     const menuIconClick = React.useCallback(() => {
-      setGridState((state) => ({
-        ...state,
-        columnMenu: { open: true, field: column.field },
-        preferencePanel: { open: false },
-      }));
-      forceUpdate();
-    }, [column.field, forceUpdate, setGridState]);
+      apiRef?.current.showColumnMenu(column.field)
+    }, [apiRef, column.field]);
 
     return (
       <div className={'MuiDataGrid-menuIcon'}>
