+++ packages/grid/_modules_/grid/components/menu/index.ts
@@ -1,2 +1,2 @@
-export * from './columnMenu/';
+export * from './columnMenu';
 export * from './GridMenu';
