+++ packages/grid/_modules_/grid/components/options-context.ts
@@ -1,4 +1,4 @@
 import * as React from 'react';
-import { DEFAULT_GRID_OPTIONS, GridOptions } from '../models';
+import { DEFAULT_GRID_OPTIONS, GridOptions } from '../models/gridOptions';
 
 export const OptionsContext = React.createContext<GridOptions>(DEFAULT_GRID_OPTIONS);
