+++ packages/grid/_modules_/grid/components/tools/StringFilterInputValueProps.ts
@@ -1,0 +1,6 @@
+import { FilterItem } from '../../models/filterItem';
+
+export interface FilterInputValueProps {
+	item: FilterItem;
+	applyValue: (value: FilterItem) => void;
+}
