+++ packages/grid/_modules_/grid/components/tools/index.ts
@@ -2,3 +2,4 @@ export * from './FilterForm';
 export * from './FilterPanel';
 export * from './Preferences';
 export * from './StringFilterInputValue';
+export { FilterInputValueProps } from './StringFilterInputValueProps';
