+++ packages/grid/_modules_/grid/hooks/features/columnMenu/columnMenuState.ts
@@ -1,0 +1,4 @@
+export interface ColumnMenuState {
+	open: boolean;
+	field?: string;
+}
