+++ packages/grid/_modules_/grid/hooks/features/filter/FilterModelState.ts
@@ -1,0 +1,11 @@
+import { FilterItem, LinkOperator } from '../../../models/filterItem';
+
+export interface FilterModelState {
+	items: FilterItem[];
+	linkOperator: LinkOperator;
+}
+
+export const getInitialFilterState: () => FilterModelState = () => ({
+	items: [],
+	linkOperator: LinkOperator.And,
+});
