+++ packages/grid/_modules_/grid/hooks/features/filter/visibleRowsState.ts
@@ -1,22 +1,4 @@
-import { FilterOperator } from '../../../models/colDef/stringColDef';
 import { RowId } from '../../../models/rows';
-
-export interface FilterItem {
-  id?: number;
-  columnField?: string;
-  value?: string;
-  operator?: FilterOperator; // Contains...
-}
-
-export enum LinkOperator {
-  And = 'and',
-  Or = 'or',
-}
-
-export interface FilterModelState {
-  items: FilterItem[];
-  linkOperator: LinkOperator;
-}
 
 export interface VisibleRowsState {
   visibleRowsLookup: Record<RowId, boolean>;
@@ -29,7 +11,3 @@ export const getInitialVisibleRowsState: () => VisibleRowsState = () => ({
   visibleRowsLookup: {},
 });
 
-export const getInitialFilterState: () => FilterModelState = () => ({
-  items: [],
-  linkOperator: LinkOperator.And,
-});
