+++ packages/grid/_modules_/grid/hooks/features/index.ts
@@ -1,13 +1,16 @@
+export * from './columnMenu';
+export * from './columnReorder';
 export * from './columns';
 export * from './core';
+export * from './filter';
+export * from './keyboard';
 export * from './pagination';
+export * from './preferencesPanel';
 export * from './rows';
 export * from './selection';
 export * from './sorting';
-export * from './keyboard';
 export * from './virtualization';
 
 export * from './useApiRef';
-export * from './useComponents';
-export * from './columnReorder';
 export * from './useColumnResize';
+export * from './useComponents';
