+++ packages/grid/_modules_/grid/hooks/features/preferencesPanel/preferencePanelSelector.ts
@@ -1,0 +1,4 @@
+import { GridState } from '../core/gridState';
+
+export const preferencePanelStateSelector = (state: GridState)=> state.preferencePanel;
+export const viewportSizeStateSelector = (state: GridState)=> state.viewportSizes;
