+++ packages/grid/_modules_/grid/hooks/features/preferencesPanel/preferencePanelState.ts
@@ -1,0 +1,7 @@
+import { PreferencePanelsValue } from './preferencesPanelValue';
+
+export interface PreferencePanelState {
+	open: boolean;
+	openedPanelValue?: PreferencePanelsValue;
+	targetField?: string;
+}
