+++ packages/grid/_modules_/grid/hooks/features/preferencesPanel/preferencesPanelValue.ts
@@ -1,0 +1,4 @@
+export enum PreferencePanelsValue {
+	filters = 'filters',
+	columns = 'columns',
+}
