+++ packages/grid/_modules_/grid/models/filterItem.ts
@@ -1,0 +1,13 @@
+import { FilterOperator } from './filterOperator';
+
+export interface FilterItem {
+	id?: number;
+	columnField?: string;
+	value?: string;
+	operator?: FilterOperator; // Contains...
+}
+
+export enum LinkOperator {
+	And = 'and',
+	Or = 'or',
+}
